Prevent Double Reduction

+++ EtaFinder.py
@@ -43,18 +43,21 @@ if not args.ft[-4:]=='npz':
     for i in range(fnames_cals.shape[0]):
         t_cals[i]=datareader.cal_time('%s/%s' %(cal_dirname,fnames_cals[i]))
     for i in range(fnames.shape[0]):
-        try:
-            print('Starting: %s' %fnames[i])
-            fname='%s/%s' %(dirname,fnames[i])
-            t_last_cal=datareader.cal_find(fname,t_cals)
-            fname_cal='%s/%s' %(cal_dirname,fnames_cals[t_cals==t_last_cal][0])
-            times, freqs,N,dynspec,temp0,template,srce=datareader.data_to_dspec(fname,fname_cal,profsig=5,sigma=10)
-            fname=fnames[i]
-            while not fname.endswith('.'):
-                fname=fname[:-1]
-            np.savez('%s/%snpz' %(dirname_save,fname),I=dynspec,freq=freqs,time=times,N=N,prof=temp0,template=template,source=srce)
-        except:
-            print('Check Calibration')
+        if not os.path.exists('%s/%snpz' %(dirname_save,%fnames[i][:-len(ftype)]))
+            try:
+                print('Starting: %s' %fnames[i])
+                fname='%s/%s' %(dirname,fnames[i])
+                t_last_cal=datareader.cal_find(fname,t_cals)
+                fname_cal='%s/%s' %(cal_dirname,fnames_cals[t_cals==t_last_cal][0])
+                times, freqs,N,dynspec,temp0,template,srce=datareader.data_to_dspec(fname,fname_cal,profsig=5,sigma=10)
+                fname=fnames[i]
+                while not fname.endswith('.'):
+                    fname=fname[:-1]
+                np.savez('%s/%snpz' %(dirname_save,fname),I=dynspec,freq=freqs,time=times,N=N,prof=temp0,template=template,source=srce)
+            except:
+                print('Check Calibration')
+        else:
+            print('File Already Exists')
 fnames=np.array([list(f for f in os.listdir(dirname_save) if f.endswith('npz'))])[0,:]
 
 times=np.zeros(fnames.shape)
